recicleView

+++ app/src/main/java/com/example/convidados/view/viewHolder/GuestViewHolder.kt
@@ -1,6 +1,8 @@
 package com.example.convidados.view.viewHolder
 
+import android.content.DialogInterface
 import android.view.View
+import androidx.appcompat.app.AlertDialog
 import androidx.recyclerview.widget.RecyclerView
 import com.example.convidados.databinding.RowGuestsBinding
 import com.example.convidados.model.GuestModel
@@ -17,7 +19,18 @@ class GuestViewHolder(private val bind: RowGuestsBinding, private val listener: 
         }
 
         bind.textName.setOnLongClickListener {
-            listener.onDelete(item.id)
+
+            AlertDialog.Builder(itemView.context)
+                .setTitle("Remoção de convidado")
+                .setMessage("Tem certeza que deseja remover?")
+                .setPositiveButton("sim"
+                ) { dialog, which ->
+                    listener.onDelete(item.id)
+                }
+                .setNegativeButton("Não", null)
+                .create().show()
+
+
             true
         }
